+++ crawl-ref/source/dungeon.h
@@ -272,8 +272,8 @@ void dgn_place_multiple_items(item_list &list,
                               const coord_def& where,
                               int level);
 
-bool set_level_flags(unsigned long flags, bool silent = false);
-bool unset_level_flags(unsigned long flags, bool silent = false);
+bool set_level_flags(uint32_t flags, bool silent = false);
+bool unset_level_flags(uint32_t flags, bool silent = false);
 
 void dgn_set_lt_callback(std::string level_type_name,
                          std::string callback_name);
